Remove static context. Add static instance to SIPTranslateMessageService.
