+++ sip-core/src/main/resources/META-INF/spring.factories
@@ -1,5 +1,4 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=de.ikor.sip.foundation.core.SIPCoreAutoConfiguration,\
-de.ikor.sip.foundation.core.util.SIPStaticSpringContext,\
 de.ikor.sip.foundation.core.translate.TranslationsAutoConfig,\
 de.ikor.sip.foundation.core.trace.TraceAutoConfig,\
 de.ikor.sip.foundation.core.proxies.ProxyAutoConfig,\
